Use pdfmark to make hyperlinks by default
I mostly make pdf output, so use pdfmark (load with `-m pdfmark` or
`-mspdf` when invoking groff`) to turn markdown links into hyperlinks.
Using footnotes as before is still supported by setting `footnote_links`
to a truthy value. Linking to headings seems possible using pdfmark but
isn't implemented here.

Similarly, images are now embedded via .PDFPIC, instead of .BP, since
the pdfpic macros are distributed with groff and I can't find where to
get mpictures from. Note that images given to be embedded need to be
pdfs themselves.

As a result of the change, it's no longer possible to style link titles
or image captions, since we're parsing everything up front. I tried
saving the title in a diversion, using .unformat on it, and then passing
it to pdfhref, but this failed an assertion in input.cpp and I couldn't
figure out how to avoid it:

    troff: src/roff/troff/input.cpp:523: static int input_stack::finish_get(node**): Assertion `level == 0' failed.

We are now handling trailing punctuation nicer, though, by eating it up
front and giving it to pdfhref as the suffix so it isn't separated from
the link by some space.